Recycle opened indexes rather than reopen
Previously an "index" was an NIO directory or a RAM directory which
could be opened into a reader/writer at will. Unfortunately getting the
requisite filesystem locks to open a new reader/writer is expensive.

Rather tha open new readers and writers, this patch creates and
maintains a global cache of (Reader, Writer, Index) tripples keyed off
of the Index or path of the Index if present. This means that multiple
threads opening a same path will get and use the same Readers and
Writers (which are thread safe) rather than attempting and failing to
open distinct Readers and Writers as the Lucene locking model only
allows for single opens.